requirements.txt
Django==1.8.4
django-allauth==0.23.0
mock==1.3.0
